requirements.txt
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.6
idna==3.7
iniconfig==2.0.0
iotopen-py-lynx==1.0.7
packaging==24.1
paho-mqtt==2.1.0
pluggy==1.5.0
pytest==8.2.2
requests==2.32.3
urllib3==2.2.2
